TadmitVideo is now independent component
